requirements.txt
# Astro Runtime includes the following pre-installed providers packages: https://www.astronomer.io/docs/astro/runtime-image-architecture#provider-packages
google-cloud-storage
google-cloud-bigquery
kaggle
kagglehub
pyyaml
astronomer-cosmos[dbt]
dbt-core
pendulum